myhealthpassport-api/requirements.txt
aerich==0.8.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioredis==2.0.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncclick==8.1.8.0
asyncpg==0.30.0
attrs==25.4.0
awscli==1.41.4
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.1.0
boto3==1.39.0
botocore==1.39.4
Brotli==1.1.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
dictdiffer==0.9.0
dnspython==2.7.0
docopt==0.6.2
docutils==0.19
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-utils==0.8.0
fonttools==4.59.1
frozenlist==1.8.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflect==7.5.0
iso8601==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.0
mypy_extensions==1.1.0
num2words==0.5.14
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.7
propcache==0.4.1
psutil==5.9.8
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
Pygments==2.19.1
pyphen==0.17.2
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.4
requirements.py==1.0.1
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.7.2
s3transfer==0.13.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlacodegen==3.2.0
SQLAlchemy==2.0.40
starlette==0.46.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli_w==1.2.0
tomlkit==0.13.2
tortoise-orm==0.24.2
typeguard==4.4.4
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
# uvloop==0.21.0  # Linux-only, not needed on Windows
watchfiles==1.0.5
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
zopfli==0.2.3.post1
